facotr out more utility functions to models from upgrades view
