+++ js-src/modules/views/upgradesView.js
@@ -7,15 +7,15 @@ var pilotsWithAbilities = require('../models/pilots').pilotsWithAbilities;
 var pilots = _.uniqBy(pilotsWithAbilities, function (pilot) {
     return pilot.text;
 });
-var upgrades = require('../models/upgrades').keyed;
 var modalController = require('../controllers/modals');
 var events = require('../controllers/events');
+var upgradesModel = require('../models/shipBuild/upgrades');
 var upgradeSlotsModel = require('../models/shipBuild/upgradeSlots');
 
 module.exports = {
-    clickEquipSlot: function (equippedUpgradesByType, filteredUpgrades, build) {
+    clickEquipSlot: function (upgradeType, equippedUpgradesByType, filteredUpgrades, build) {
         // open modal to choose upgrade to equip
-        var $modalContent = module.exports.renderUpgradeModalContent(build, filteredUpgrades, equippedUpgradesByType);
+        var $modalContent = module.exports.renderUpgradeModalContent(build, upgradeType, filteredUpgrades, equippedUpgradesByType);
         modalController.openOptionSelectModal($modalContent, 'Buy upgrade');
     },
     removeEquipSlot: function (upgradeId) {
@@ -23,8 +23,7 @@ module.exports = {
     },
     renderUpgradesList: function (build) {
         // Get a list of the slots allowed for this build (determined by ship and PS) and the number of each upgrade per slot
-        var upgradesAllowedInBuild = module.exports.numberOfUsableUpgrades(build.pilotSkill, build.currentShip, build.upgrades);
-        var upgrades = module.exports.getAllowedUpgrades(build, upgradesAllowedInBuild);
+        var upgrades = upgradesModel.getAllowedUpgrades(build);
 
         var $freeList = $('#free-upgrade-list');
         $freeList.empty();
@@ -70,29 +69,6 @@ module.exports = {
             $('.disabled-upgrades').hide();
         }
     },
-    getAllowedUpgrades: function (build, upgradesAllowedInBuild) {
-        var purchasedUpgradesByType = _.clone(build.upgrades, true);
-
-        var allowedUpgrades = [];
-        var disallowedUpgrades = [];
-
-        // Populate allowed and disallowed slot lists with the purchased upgrades
-        _.forEach(purchasedUpgradesByType, function (upgradesList, slotType) {
-            // Find the keys of any slots in the build that match these upgrades
-            if (upgradesAllowedInBuild[slotType]) {
-                // Allowed in build
-                allowedUpgrades = allowedUpgrades.concat(upgradesList);
-            } else {
-                // This slot not allowed in this build (restricted by ship?)
-                disallowedUpgrades = disallowedUpgrades.concat(upgradesList);
-            }
-        });
-
-        return {
-            allowed: allowedUpgrades,
-            disallowed: disallowedUpgrades
-        };
-    },
     renderUpgradeItem: function (upgrade) {
         var imageUrl = '/components/xwing-data/images/' + upgrade.image;
         var $item = $('<li class="upgrade" data-featherlight="' + imageUrl + '">' + module.exports.getIconString(upgrade.slot) + '<span>' + upgrade.name + '</span><i class="material-icons eye">remove_red_eye</i><img class="preview" src="' + imageUrl + '"></li>');
@@ -110,50 +86,32 @@ module.exports = {
         var iconString = '<i class="xwing-miniatures-font xwing-miniatures-font-' + iconId + '"></i>';
         return iconString;
     },
-    getFilteredUpgrades: function (upgradeTypeString, existingUpgrades, currentShip) {
-        var upgradeTypes = upgradeTypeString.split(',');
-
-        var filteredUpgradesByType = {};
-
-        _.each(upgradeTypes, function (upgradeType) {
-            var upgradesOfType = upgrades[upgradeType];
-            var existingUpgradesOfType = existingUpgrades[upgradeType];
-            var filteredUpgrades = module.exports.getUpgradesToShow(upgradesOfType, currentShip, existingUpgradesOfType);
-            if (filteredUpgrades.length > 0) {
-                filteredUpgradesByType[upgradeType] = filteredUpgrades;
-            }
-        });
-
-        return filteredUpgradesByType;
-    },
-    renderUpgradeModalContent: function (build, filteredUpgradesByType, equippedUpgrades) {
+    renderUpgradeModalContent: function (build, upgradeType, filteredUpgradesByType, equippedUpgrades) {
         var tabs = [];
 
-        _.each(filteredUpgradesByType, function (filteredUpgrades, upgradeType) {
-            var $equippedUpgradesTab = module.exports.renderEquippedCardListModalContent(upgradeType, build, equippedUpgrades);
-            tabs.push({
-                name: 'Existing ' + upgradeType,
-                $content: $equippedUpgradesTab
-            });
-
-            var $tab = module.exports.renderCardListModalContent(upgradeType, build, filteredUpgrades);
-            var tabName = 'Buy new ' + upgradeType;
-            if (upgradeType === 'Elite') {
-                tabName = 'Elite cards';
-            }
-            tabs.push({
-                name: tabName,
-                $content: $tab
-            });
-
-            if (upgradeType === 'Elite') {
-                var $abilityTab = module.exports.renderPilotAbilityModalContent(build);
-                tabs.push({
-                    name: 'Pilot abilities',
-                    $content: $abilityTab
-                });
-            }
+        var $equippedUpgradesTab = module.exports.renderEquippedCardListModalContent(upgradeType, build, equippedUpgrades);
+        tabs.push({
+            name: 'Existing ' + upgradeType,
+            $content: $equippedUpgradesTab
         });
+
+        var $tab = module.exports.renderCardListModalContent(upgradeType, build, filteredUpgradesByType);
+        var tabName = 'Buy new ' + upgradeType;
+        if (upgradeType === 'Elite') {
+            tabName = 'Elite cards';
+        }
+        tabs.push({
+            name: tabName,
+            $content: $tab
+        });
+
+        if (upgradeType === 'Elite') {
+            var $abilityTab = module.exports.renderPilotAbilityModalContent(build);
+            tabs.push({
+                name: 'Pilot abilities',
+                $content: $abilityTab
+            });
+        }
 
         return module.exports.renderTabs(tabs);
     },
@@ -235,54 +193,6 @@ module.exports = {
 
         return $modalContent;
     },
-    getUpgradesToShow: function (upgrades, currentShip, existingUpgrades) {
-        var filteredUpgrades = _.filter(upgrades, function (upgrade) {
-            // Remove any upgrades for different ships
-            if (upgrade.ship && upgrade.ship.indexOf(currentShip.shipData.name) < 0) {
-                return false;
-            }
-
-            // Remove any upgrades for different ship sizes
-            if (upgrade.size && upgrade.size.indexOf(currentShip.shipData.size) < 0) {
-                return false;
-            }
-
-            // Don't show anything which is a starting upgrade for the ship
-            if (currentShip.startingUpgrades) {
-                var found = _.find(currentShip.startingUpgrades, function (startingUpgrade) {
-                    return startingUpgrade.xws === upgrade.xws;
-                });
-                if (found) {
-                    return false;
-                }
-            }
-
-            // Remove any upgrades the build already has
-            var upgradeExists = _.find(existingUpgrades, function (existingUpgrade) {
-                return existingUpgrade.id === upgrade.id;
-            });
-
-            if (upgradeExists) {
-                var upgradeIsAllowed = false;
-                // filter out any upgrades the player already has
-                // except
-                // * secondary weapons & bombs
-                if (upgrade.slot === 'Bomb' || upgrade.slot === 'Torpedo' || upgrade.slot === 'Cannon' || upgrade.slot === 'Turret' || upgrade.slot === 'Missile') {
-                    upgradeIsAllowed = true;
-                // * hull upgrade and shield upgrade
-                } else if (upgrade.xws === 'hullupgrade' || upgrade.xws === 'shieldupgrade') {
-                    upgradeIsAllowed = true;
-                }
-                if (!upgradeIsAllowed) {
-                    return false;
-                }
-            }
-
-            return true;
-        });
-
-        return filteredUpgrades;
-    },
     renderPilotAbilityModalContent: function (build) {
         var $modalContent = $('<div class="pilot-ability-list" id="modal-pilot-ability-list">');
         var $upgradeList = $('<ul>');
@@ -317,73 +227,6 @@ module.exports = {
 
         return $modalContent;
     },
-    numberOfUsableUpgrades: function (pilotSkill, currentShip, upgradesByType) {
-
-        // elite slots are dependent on pilot level
-        var eliteSlots = 0;
-        if (pilotSkill >= 9) {
-            eliteSlots = 4;
-        } else if (pilotSkill >= 7) {
-            eliteSlots = 3;
-        } else if (pilotSkill >= 5) {
-            eliteSlots = 2;
-        } else if (pilotSkill >= 3) {
-            eliteSlots = 1;
-        }
-
-        // mod slots are dependent on pilot level
-        var modSlots = 1;
-        if (pilotSkill >= 8) {
-            modSlots = 4;
-        } else if (pilotSkill >= 6) {
-            modSlots = 3;
-        } else if (pilotSkill >= 4) {
-            modSlots = 2;
-        }
-
-        var usableUpgrades = {
-            Modification: {
-                max: 3,
-                allowed: modSlots
-            },
-            Title: {
-                max: 1,
-                allowed: 1
-            }
-        };
-        if (eliteSlots > 0) {
-            usableUpgrades.Elite = {
-                max: 4,
-                allowed: eliteSlots
-            };
-        }
-
-        var addToSlot = function (slotKey, max, allowed) {
-            if (!usableUpgrades[slotKey]) {
-                usableUpgrades[slotKey] = {
-                    max: max,
-                    allowed: allowed
-                };
-            } else {
-                usableUpgrades[slotKey].max += max;
-                usableUpgrades[slotKey].allowed += allowed;
-            }
-        };
-
-        // Add slots for the ship type
-        var upgradeSlots = currentShip.upgradeSlots;
-        _.each(upgradeSlots, function (slotType) {
-            addToSlot(slotType, 1, 1);
-        });
-
-        // Add slots for the upgrade cards
-        var slotsFromUpgrades = upgradeSlotsModel.getSlotsFromUpgrades(usableUpgrades, currentShip.startingUpgrades, upgradesByType);
-        _.each(slotsFromUpgrades, function (slotType) {
-            addToSlot(slotType, 1, 1);
-        });
-
-        return usableUpgrades;
-    },
     renderShipUpgrades: function (build) {
         // Process and create list for ship chassis slots
         var $shipSlots = $('#ship-slots-default');
@@ -397,10 +240,10 @@ module.exports = {
         // var equippedAbilities = $.extend(true, [], build.equippedUpgrades.pilotAbilities);
 
         _.each(upgradeSlots, function (upgradeSlot) {
-            var filteredUpgradesByType = module.exports.getFilteredUpgrades(upgradeSlot.type, build.upgrades, build.currentShip);
+            var filteredUpgradesByType = upgradesModel.getFilteredUpgrades(upgradeSlot.type, build.upgrades, build.currentShip);
 
             // Don't show this slot if there are no available upgrades for it (e.g. a title slot for a ship with no titles)
-            if (!filteredUpgradesByType[upgradeSlot.type] || filteredUpgradesByType[upgradeSlot.type].length < 1) {
+            if (!filteredUpgradesByType || filteredUpgradesByType.length < 1) {
                 return;
             }
 
@@ -452,7 +295,7 @@ module.exports = {
                     });
                 } else {
                     $slot.on('click', function () {
-                        module.exports.clickEquipSlot(build.upgrades[upgradeSlot.type], filteredUpgradesByType, build);
+                        module.exports.clickEquipSlot(upgradeSlot.type, build.upgrades[upgradeSlot.type], filteredUpgradesByType, build);
                     });
                 }
